+++ packages/vite-plugin-remix-mui/example/app/moo-patch-2.ts
@@ -1,7 +1,0 @@
-import * as moo from "./moo";
-
-export default function () {
-  console.log("moo 2 start");
-  moo.default();
-  console.log("moo 2 end");
-}