+++ packages/vite-plugin-remix-mui/example/vite.config.ts
@@ -1,6 +1,4 @@
 import mui from "@postinumero/vite-plugin-remix-mui";
-import moduleProxy from "@postinumero/vite-plugin-replace-module";
-import remixRoot from "@postinumero/vite-plugin-replace-module/lib/presets/remix-root.js";
 import { vitePlugin as remix } from "@remix-run/dev";
 import { VitePluginConfig } from "@remix-run/dev/dist/vite/plugin";
 import { defineConfig } from "vite";
@@ -9,18 +7,10 @@ import tsconfigPaths from "vite-tsconfig-paths";
 export const config = async (
   options: {
     remix?: VitePluginConfig;
-  } = {}
+  } = {},
 ) =>
   defineConfig({
-    plugins: [
-      remix(options.remix),
-      tsconfigPaths(),
-      mui,
-      ...(await remixRoot({ url: import.meta.url })),
-      moduleProxy({ id: "./moo", proxy: "./app/moo-patch-1.ts" }),
-      moduleProxy({ id: "./moo", proxy: "./app/moo-patch-2.ts" }),
-      moduleProxy({ id: "./moo", proxy: "./app/moo-patch-3.ts" }),
-    ],
+    plugins: [remix(options.remix), tsconfigPaths(), mui],
   });
 
 export default config();